# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "twitter-as-a-database",
    "description": "Using Twitter as a Database as a hobby project!",
    "externalDocs": {
      "description": "GitHub Repository",
      "url": "https://github.com/syall/Twitter-as-a-Database"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "development"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Returns a list of Public Users",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "A list of Public Users.",
            "content": {
              "application/json": {
                "examples": {
                  "steven": {
                    "value": [
                      {
                        "user": "steven",
                        "active": "true"
                      }
                    ]
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Public Users"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "message": "Unable to get Public Users."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/search/{query}": {
      "get": {
        "summary": "Returns a list of Public Users that start with the Query",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "String to Query"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Public Users that start with the Query.",
            "content": {
              "application/json": {
                "examples": {
                  "steven": {
                    "value": [
                      {
                        "user": "steven",
                        "active": "true"
                      }
                    ]
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Public Users"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "message": "Unable to get Public Users."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/active": {
      "get": {
        "summary": "Returns a list of Public Users that are active",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "A list of Public Users that are active.",
            "content": {
              "application/json": {
                "examples": {
                  "steven": {
                    "value": [
                      {
                        "user": "steven",
                        "active": "true"
                      }
                    ]
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Public Users"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "message": "Unable to get Public Users."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/create": {
      "post": {
        "summary": "Returns Public User created",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "description": "Creation Information",
          "required": true,
          "content": {
            "application/json": {
              "examples": {
                "bob": {
                  "value": {
                    "user": "bob",
                    "password": "test"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/Creation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Public User created.",
            "content": {
              "application/json": {
                "examples": {
                  "bob": {
                    "value": {
                      "user": "bob",
                      "active": "true"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Public User"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "message": "Unable to create user."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/read/{user}": {
      "get": {
        "summary": "Returns Public User",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "User to get"
          }
        ],
        "responses": {
          "200": {
            "description": "Public User requested.",
            "content": {
              "application/json": {
                "examples": {
                  "steven": {
                    "value": {
                      "user": "steven",
                      "active": "true"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Public User"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "message": "Unable to get user."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/update/{user}": {
      "put": {
        "summary": "Returns updated Public User.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "User to update"
          }
        ],
        "responses": {
          "200": {
            "description": "Public User updated.",
            "content": {
              "application/json": {
                "examples": {
                  "steven": {
                    "value": {
                      "user": "steven",
                      "active": "true"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Public User"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "message": "Unable to update user."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/delete/{user}": {
      "delete": {
        "summary": "Returns deleted Public User",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "User to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Public User deleted.",
            "content": {
              "application/json": {
                "examples": {
                  "steven": {
                    "value": {
                      "user": "steven",
                      "active": "true"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Public User"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "message": "Unable to delete user."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticates a User",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "description": "Login Information",
          "required": true,
          "content": {
            "application/json": {
              "examples": {
                "steven": {
                  "value": {
                    "username": "steven",
                    "password": "test123"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication Successful!",
            "content": {
              "application/json": {
                "examples": {
                  "steven": {
                    "value": {
                      "message": "Authentication Successful!"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Login Success"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "examples": {
                  "userNotFound": {
                    "value": {
                      "message": "User not found."
                    }
                  },
                  "passwordIncorrect": {
                    "value": {
                      "message": "Password Incorrect."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "username",
          "password",
          "active"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "active": {
            "type": "active"
          }
        }
      },
      "Public User": {
        "type": "object",
        "required": [
          "username",
          "active"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "Public Users": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Public User"
        }
      },
      "Login": {
        "type": "object",
        "require": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Login Success": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Creation": {
        "type": "object",
        "require": [
          "user",
          "password"
        ],
        "properties": {
          "user": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    }
  }
}